feat(media): decompose media pages and extract shared components (14 issues)
Closes #1990 — remove duplicate formatRuntime (lib/utils.ts + MovieSearchResult)
Closes #1991 — remove duplicate formatEpisodeCode from CalendarPage
Closes #1994 — replace ProgressBar internals with @pops/ui Progress primitive
Closes #1997 — create lib/statusStyles.ts; consolidate badge colour maps
Closes #2049 — create CardWithActionOverlay; refactor ComparisonMovieCard + DiscoverCard
Closes #2050 — extract ExpandableListRow from EpisodeList
Closes #2051 — create ResponsiveActionBar (canonical rename of ArenaActionBar)
Closes #2055 — create ConditionalModalButton from RequestMovieButton pattern
Closes #2056 — create CRUDManagementSection; refactor SourceManagementSection + DimensionManager
Closes #2057 — extract SummaryCard from DebriefControls
Closes #2059 — remove unused DebriefComparisonCard (superseded by ComparisonMovieCard)
Closes #2040 — extract WatchlistCard + DrawTierButtons sub-components
Closes #2046 — create RequestDialog base; update RequestMovieModal + RequestSeriesModal
Closes #2062 — lift SearchInput + MobileSearchOverlay + searchStore to @pops/navigation
