+++ packages/app-media/src/components/ArrStatusBadge.tsx
@@ -6,6 +6,8 @@ import { trpc } from '@pops/api-client';
  */
 import { Badge } from '@pops/ui';
 
+import { ARR_STATUS_STYLES } from '../lib/statusStyles';
+
 type MediaKind = 'movie' | 'show';
 
 interface ArrStatusBadgeProps {
@@ -13,16 +15,6 @@ interface ArrStatusBadgeProps {
   /** TMDB ID for movies, TVDB ID for shows. */
   externalId: number;
 }
-
-const STATUS_STYLES: Record<string, string> = {
-  available: 'bg-success text-success-foreground',
-  complete: 'bg-success text-success-foreground',
-  monitored: 'bg-warning text-warning-foreground',
-  downloading: 'bg-warning text-warning-foreground',
-  partial: 'bg-warning text-warning-foreground',
-  unmonitored: 'bg-muted text-muted-foreground',
-  not_found: 'bg-muted text-muted-foreground',
-};
 
 export function ArrStatusBadge({ kind, externalId }: ArrStatusBadgeProps) {
   const { data: configData } = trpc.media.arr.getConfig.useQuery();
@@ -54,7 +46,7 @@ export function ArrStatusBadge({ kind, externalId }: ArrStatusBadgeProps) {
   if (!query.data?.data) return null;
 
   const result = query.data.data;
-  const colorClass = STATUS_STYLES[result.status] ?? STATUS_STYLES.not_found;
+  const colorClass = ARR_STATUS_STYLES[result.status] ?? ARR_STATUS_STYLES.not_found;
 
   return <Badge className={colorClass}>{result.label}</Badge>;
 }
