+++ packages/app-media/src/components/ExpandableListRow.tsx
@@ -1,0 +1,148 @@
+import { Check, ChevronDown, ChevronRight, HardDrive } from 'lucide-react';
+import { useState } from 'react';
+
+import { Switch } from '@pops/ui';
+
+import { formatRuntime } from '../lib/format';
+
+/**
+ * ExpandableListRow — a single row in a bordered list with optional:
+ *   - watched checkbox
+ *   - expand/collapse chevron for overview text
+ *   - air-date + runtime metadata
+ *   - downloaded (HardDrive) indicator
+ *   - Sonarr monitoring toggle
+ *
+ * Extracted from EpisodeList so it can be reused in other list contexts.
+ */
+
+export interface ExpandableListRowItem {
+  id: number;
+  episodeNumber: number;
+  name: string | null;
+  overview: string | null;
+  airDate: string | null;
+  runtime: number | null;
+}
+
+export interface ExpandableListRowProps {
+  item: ExpandableListRowItem;
+  isWatched?: boolean;
+  isToggling?: boolean;
+  onToggleWatched?: (id: number, watched: boolean) => void;
+  isMonitored?: boolean;
+  hasFile?: boolean;
+  onToggleMonitored?: (episodeNumber: number, monitored: boolean) => void;
+  isMonitoringPending?: boolean;
+  isUpcoming?: boolean;
+}
+
+export function ExpandableListRow({
+  item,
+  isWatched = false,
+  isToggling = false,
+  onToggleWatched,
+  isMonitored,
+  hasFile,
+  onToggleMonitored,
+  isMonitoringPending,
+  isUpcoming = false,
+}: ExpandableListRowProps) {
+  const [isExpanded, setIsExpanded] = useState(false);
+  const hasOverview = !!item.overview;
+
+  return (
+    <div className={`px-4 py-3${isUpcoming ? ' opacity-50' : ''}`}>
+      <div className="flex items-start gap-3">
+        {onToggleWatched && (
+          <button
+            type="button"
+            onClick={() => {
+              onToggleWatched(item.id, !isWatched);
+            }}
+            disabled={isToggling || isUpcoming}
+            aria-label={
+              isUpcoming
+                ? `Episode ${item.episodeNumber} upcoming`
+                : isWatched
+                  ? `Mark episode ${item.episodeNumber} as unwatched`
+                  : `Mark episode ${item.episodeNumber} as watched`
+            }
+            className={`mt-0.5 shrink-0 flex items-center justify-center h-5 w-5 rounded border transition-colors ${
+              isToggling || isUpcoming
+                ? 'opacity-50 cursor-not-allowed border-muted'
+                : isWatched
+                  ? 'bg-primary border-primary text-primary-foreground'
+                  : 'border-muted-foreground/40 hover:border-primary'
+            }`}
+          >
+            {isWatched && <Check className="h-3.5 w-3.5" />}
+          </button>
+        )}
+
+        {hasOverview ? (
+          <button
+            type="button"
+            onClick={() => setIsExpanded((v) => !v)}
+            aria-expanded={isExpanded}
+            className="mt-0.5 text-muted-foreground shrink-0 hover:text-foreground"
+          >
+            {isExpanded ? (
+              <ChevronDown className="h-4 w-4" />
+            ) : (
+              <ChevronRight className="h-4 w-4" />
+            )}
+          </button>
+        ) : null}
+
+        <div className="flex-1 min-w-0">
+          <div className="flex items-baseline gap-2">
+            <span className="text-sm font-medium text-muted-foreground shrink-0">
+              {item.episodeNumber}
+            </span>
+            <span
+              className={`text-sm font-medium truncate ${isWatched ? 'text-muted-foreground' : ''}`}
+            >
+              {item.name ?? `Episode ${item.episodeNumber}`}
+            </span>
+            {isUpcoming && (
+              <span className="text-xs text-warning font-medium shrink-0">Upcoming</span>
+            )}
+          </div>
+
+          <div className="flex items-center gap-2 mt-0.5 text-xs text-muted-foreground">
+            {item.airDate && <span>{item.airDate}</span>}
+            {item.airDate && item.runtime && <span>·</span>}
+            {item.runtime && <span>{formatRuntime(item.runtime)}</span>}
+          </div>
+        </div>
+
+        {hasFile && (
+          <span
+            className="shrink-0 text-success"
+            title="Downloaded"
+            aria-label={`Episode ${item.episodeNumber} downloaded`}
+          >
+            <HardDrive className="h-4 w-4" />
+          </span>
+        )}
+
+        {onToggleMonitored && isMonitored !== undefined && (
+          <Switch
+            size="sm"
+            checked={isMonitored}
+            aria-label={`Monitor episode ${item.episodeNumber}`}
+            disabled={isMonitoringPending}
+            onCheckedChange={(checked: boolean) => {
+              onToggleMonitored(item.episodeNumber, checked);
+            }}
+          />
+        )}
+      </div>
+
+      {isExpanded && hasOverview && (
+        <p className="mt-2 ml-7 text-sm text-muted-foreground leading-relaxed">{item.overview}</p>
+      )}
+    </div>
+  );
+}
