+++ packages/app-media/src/components/RequestMovieButton.tsx
@@ -1,5 +1,4 @@
 import { Download } from 'lucide-react';
-import { useState } from 'react';
 
 import { trpc } from '@pops/api-client';
 /**
@@ -8,9 +7,12 @@ import { trpc } from '@pops/api-client';
  * Hidden when the movie already exists in Radarr.
  * Disabled when Radarr is not configured.
  * Returns null on query error (service unreachable).
+ *
+ * Uses ConditionalModalButton to manage the RequestMovieModal open state.
  */
 import { Button } from '@pops/ui';
 
+import { ConditionalModalButton } from './ConditionalModalButton';
 import { RequestMovieModal } from './RequestMovieModal';
 
 type ButtonVariant = 'standard' | 'compact';
@@ -30,8 +32,6 @@ export function RequestMovieButton({
   variant = 'standard',
   onRequest,
 }: RequestMovieButtonProps) {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
   const { data: configData } = trpc.media.arr.getConfig.useQuery();
   const config = configData?.data;
 
@@ -64,7 +64,7 @@ export function RequestMovieButton({
     );
   }
 
-  // Loading — show spinner
+  // Loading — show nothing
   if (movieStatus.isLoading) return null;
 
   // Error — hide entirely (service unreachable)
@@ -74,59 +74,55 @@ export function RequestMovieButton({
   const status = movieStatus.data?.data?.status;
   if (status && status !== 'not_found') return null;
 
-  const handleClick = () => {
-    if (onRequest) {
-      onRequest(tmdbId);
-    } else {
-      setIsModalOpen(true);
-    }
-  };
-
-  if (variant === 'compact') {
-    return (
-      <>
+  // onRequest override — caller handles the action directly
+  if (onRequest) {
+    if (variant === 'compact') {
+      return (
         <Button
           size="icon"
           variant="ghost"
           className="h-7 w-7 text-white hover:bg-white/20"
-          onClick={handleClick}
+          onClick={() => onRequest(tmdbId)}
           title="Request in Radarr"
           aria-label="Request in Radarr"
         >
           <Download className="h-3.5 w-3.5" />
         </Button>
-        {!onRequest && (
-          <RequestMovieModal
-            open={isModalOpen}
-            onClose={() => {
-              setIsModalOpen(false);
-            }}
-            tmdbId={tmdbId}
-            title={title}
-            year={year}
-          />
-        )}
-      </>
+      );
+    }
+    return (
+      <Button variant="outline" size="sm" onClick={() => onRequest(tmdbId)}>
+        <Download className="h-4 w-4" />
+        Request
+      </Button>
     );
   }
 
   return (
-    <>
-      <Button variant="outline" size="sm" onClick={handleClick}>
-        <Download className="h-4 w-4" />
-        Request
-      </Button>
-      {!onRequest && (
-        <RequestMovieModal
-          open={isModalOpen}
-          onClose={() => {
-            setIsModalOpen(false);
-          }}
-          tmdbId={tmdbId}
-          title={title}
-          year={year}
-        />
+    <ConditionalModalButton
+      show
+      trigger={({ onClick }) =>
+        variant === 'compact' ? (
+          <Button
+            size="icon"
+            variant="ghost"
+            className="h-7 w-7 text-white hover:bg-white/20"
+            onClick={onClick}
+            title="Request in Radarr"
+            aria-label="Request in Radarr"
+          >
+            <Download className="h-3.5 w-3.5" />
+          </Button>
+        ) : (
+          <Button variant="outline" size="sm" onClick={onClick}>
+            <Download className="h-4 w-4" />
+            Request
+          </Button>
+        )
+      }
+      modal={({ open, onClose }) => (
+        <RequestMovieModal open={open} onClose={onClose} tmdbId={tmdbId} title={title} year={year} />
       )}
-    </>
+    />
   );
 }
