+++ packages/app-media/src/components/RequestSeriesModal.tsx
@@ -1,4 +1,3 @@
-import { CheckCircle2, RefreshCw } from 'lucide-react';
 import { useEffect, useMemo, useState } from 'react';
 
 import { trpc } from '@pops/api-client';
@@ -8,17 +7,9 @@ import { trpc } from '@pops/api-client';
  * Presents quality profile, root folder, and language profile dropdowns,
  * season monitoring checkboxes with smart defaults, then submits to Sonarr.
  */
-import {
-  Button,
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  formatBytes,
-  Label,
-  Select,
-} from '@pops/ui';
+import { Button, formatBytes, Label, Select } from '@pops/ui';
+
+import { RequestDialog } from './RequestDialog';
 
 export interface SeasonInfo {
   seasonNumber: number;
@@ -35,9 +26,8 @@ interface RequestSeriesModalProps {
   seasons: SeasonInfo[];
 }
 
-/** Determine if a season is "future" (should be monitored by default). */
 function isFutureSeason(firstAirDate: string | null): boolean {
-  if (!firstAirDate) return true; // unannounced = future
+  if (!firstAirDate) return true;
   return new Date(firstAirDate) > new Date();
 }
 
@@ -69,7 +59,6 @@ export function RequestSeriesModal({
     retry: false,
   });
 
-  // Default to first option once loaded
   useEffect(() => {
     if (profiles.data?.data?.length && qualityProfileId === null) {
       setQualityProfileId(profiles.data.data[0]!.id);
@@ -88,7 +77,6 @@ export function RequestSeriesModal({
     }
   }, [languages.data?.data, languageProfileId]);
 
-  // Set smart season defaults when seasons change
   useEffect(() => {
     if (seasons.length > 0 && Object.keys(seasonMonitored).length === 0) {
       const defaults: Record<number, boolean> = {};
@@ -135,7 +123,7 @@ export function RequestSeriesModal({
   const isDataLoading = profiles.isLoading || folders.isLoading || languages.isLoading;
   const canSubmit =
     qualityProfileId !== null &&
-    rootFolderPath &&
+    !!rootFolderPath &&
     languageProfileId !== null &&
     !addSeries.isPending &&
     !success;
@@ -152,202 +140,169 @@ export function RequestSeriesModal({
     [seasons, seasonMonitored]
   );
 
-  return (
-    <Dialog open={open} onOpenChange={(v) => !v && handleClose()}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Request Series</DialogTitle>
-          <DialogDescription>
-            {title} ({year})
-          </DialogDescription>
-        </DialogHeader>
+  const handleSubmit = () => {
+    if (qualityProfileId !== null && rootFolderPath && languageProfileId !== null) {
+      setError(null);
+      addSeries.mutate({
+        tvdbId,
+        title,
+        qualityProfileId,
+        rootFolderPath,
+        languageProfileId,
+        seasons: seasons.map((s) => ({
+          seasonNumber: s.seasonNumber,
+          monitored: seasonMonitored[s.seasonNumber] ?? false,
+        })),
+      });
+    }
+  };
 
-        <div className="space-y-4 pt-2">
-          {isDataLoading ? (
-            <div className="flex items-center justify-center py-8 text-sm text-muted-foreground">
-              <RefreshCw className="h-4 w-4 mr-2 animate-spin" />
-              Loading options...
-            </div>
-          ) : profileList.length === 0 || folderList.length === 0 || languageList.length === 0 ? (
-            <div className="text-center py-4 space-y-2">
-              <p className="text-sm text-destructive/80">
-                {profileList.length === 0
-                  ? 'No quality profiles found'
-                  : folderList.length === 0
-                    ? 'No root folders found'
-                    : 'No language profiles found'}
-                .
-              </p>
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => {
-                  profiles.refetch();
-                  folders.refetch();
-                  languages.refetch();
-                }}
-              >
-                Retry
-              </Button>
-            </div>
-          ) : (
-            <>
-              {/* Quality Profile */}
-              <Select
-                label="Quality Profile"
-                id="quality-profile"
-                value={String(qualityProfileId ?? '')}
-                onChange={(e) => {
-                  setQualityProfileId(Number(e.target.value));
-                }}
-                disabled={addSeries.isPending || success}
-                options={profileList.map((p) => ({
-                  value: String(p.id),
-                  label: p.name,
-                }))}
-              />
+  const formContent =
+    profileList.length === 0 || folderList.length === 0 || languageList.length === 0 ? (
+      <div className="text-center py-4 space-y-2">
+        <p className="text-sm text-destructive/80">
+          {profileList.length === 0
+            ? 'No quality profiles found'
+            : folderList.length === 0
+              ? 'No root folders found'
+              : 'No language profiles found'}
+          .
+        </p>
+        <Button
+          variant="outline"
+          size="sm"
+          onClick={() => {
+            profiles.refetch();
+            folders.refetch();
+            languages.refetch();
+          }}
+        >
+          Retry
+        </Button>
+      </div>
+    ) : (
+      <>
+        <Select
+          label="Quality Profile"
+          id="quality-profile"
+          value={String(qualityProfileId ?? '')}
+          onChange={(e) => {
+            setQualityProfileId(Number(e.target.value));
+          }}
+          disabled={addSeries.isPending || success}
+          options={profileList.map((p) => ({
+            value: String(p.id),
+            label: p.name,
+          }))}
+        />
 
-              {/* Root Folder */}
-              <Select
-                label="Root Folder"
-                id="root-folder"
-                value={rootFolderPath}
-                onChange={(e) => {
-                  setRootFolderPath(e.target.value);
-                }}
-                disabled={addSeries.isPending || success}
-                options={folderList.map((f) => ({
-                  value: f.path,
-                  label: `${f.path} (${formatBytes(f.freeSpace)} free)`,
-                }))}
-              />
+        <Select
+          label="Root Folder"
+          id="root-folder"
+          value={rootFolderPath}
+          onChange={(e) => {
+            setRootFolderPath(e.target.value);
+          }}
+          disabled={addSeries.isPending || success}
+          options={folderList.map((f) => ({
+            value: f.path,
+            label: `${f.path} (${formatBytes(f.freeSpace)} free)`,
+          }))}
+        />
 
-              {/* Language Profile */}
-              <Select
-                label="Language Profile"
-                id="language-profile"
-                value={String(languageProfileId ?? '')}
-                onChange={(e) => {
-                  setLanguageProfileId(Number(e.target.value));
-                }}
-                disabled={addSeries.isPending || success}
-                options={languageList.map((l) => ({
-                  value: String(l.id),
-                  label: l.name,
-                }))}
-              />
+        <Select
+          label="Language Profile"
+          id="language-profile"
+          value={String(languageProfileId ?? '')}
+          onChange={(e) => {
+            setLanguageProfileId(Number(e.target.value));
+          }}
+          disabled={addSeries.isPending || success}
+          options={languageList.map((l) => ({
+            value: String(l.id),
+            label: l.name,
+          }))}
+        />
 
-              {/* Season Monitoring */}
-              {seasons.length > 0 && (
-                <div className="space-y-1.5">
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm font-medium">Season Monitoring</span>
-                    {showBulkControls && (
-                      <div className="flex gap-2">
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          className="text-xs h-auto p-0 text-muted-foreground hover:text-foreground"
-                          disabled={allChecked || addSeries.isPending || success}
-                          onClick={() => {
-                            const all: Record<number, boolean> = {};
-                            for (const s of seasons) all[s.seasonNumber] = true;
-                            setSeasonMonitored(all);
-                          }}
-                        >
-                          Select All
-                        </Button>
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          className="text-xs h-auto p-0 text-muted-foreground hover:text-foreground"
-                          disabled={noneChecked || addSeries.isPending || success}
-                          onClick={() => {
-                            const none: Record<number, boolean> = {};
-                            for (const s of seasons) none[s.seasonNumber] = false;
-                            setSeasonMonitored(none);
-                          }}
-                        >
-                          Deselect All
-                        </Button>
-                      </div>
-                    )}
-                  </div>
-                  <div className="max-h-48 overflow-y-auto space-y-1 rounded-md border p-2">
-                    {seasons.map((s) => (
-                      <Label
-                        key={s.seasonNumber}
-                        className="flex items-center gap-2 text-sm cursor-pointer font-normal"
-                      >
-                        <input
-                          type="checkbox"
-                          checked={seasonMonitored[s.seasonNumber] ?? false}
-                          onChange={(e) => {
-                            setSeasonMonitored((prev) => ({
-                              ...prev,
-                              [s.seasonNumber]: e.target.checked,
-                            }));
-                          }}
-                          disabled={addSeries.isPending || success}
-                        />
-                        {s.seasonNumber === 0 ? 'Specials' : `Season ${s.seasonNumber}`}
-                        {s.firstAirDate && (
-                          <span className="text-muted-foreground">
-                            — {s.firstAirDate.slice(0, 4)}
-                          </span>
-                        )}
-                      </Label>
-                    ))}
-                  </div>
+        {seasons.length > 0 && (
+          <div className="space-y-1.5">
+            <div className="flex items-center justify-between">
+              <span className="text-sm font-medium">Season Monitoring</span>
+              {showBulkControls && (
+                <div className="flex gap-2">
+                  <Button
+                    variant="ghost"
+                    size="sm"
+                    className="text-xs h-auto p-0 text-muted-foreground hover:text-foreground"
+                    disabled={allChecked || addSeries.isPending || success}
+                    onClick={() => {
+                      const all: Record<number, boolean> = {};
+                      for (const s of seasons) all[s.seasonNumber] = true;
+                      setSeasonMonitored(all);
+                    }}
+                  >
+                    Select All
+                  </Button>
+                  <Button
+                    variant="ghost"
+                    size="sm"
+                    className="text-xs h-auto p-0 text-muted-foreground hover:text-foreground"
+                    disabled={noneChecked || addSeries.isPending || success}
+                    onClick={() => {
+                      const none: Record<number, boolean> = {};
+                      for (const s of seasons) none[s.seasonNumber] = false;
+                      setSeasonMonitored(none);
+                    }}
+                  >
+                    Deselect All
+                  </Button>
                 </div>
               )}
-            </>
-          )}
-
-          {/* Error */}
-          {error && <p className="text-sm text-destructive/80">{error}</p>}
-
-          {/* Actions */}
-          <div className="flex justify-end gap-2 pt-2">
-            <Button variant="outline" onClick={handleClose} disabled={addSeries.isPending}>
-              Cancel
-            </Button>
-            <Button
-              onClick={() => {
-                if (qualityProfileId !== null && rootFolderPath && languageProfileId !== null) {
-                  setError(null);
-                  addSeries.mutate({
-                    tvdbId,
-                    title,
-                    qualityProfileId,
-                    rootFolderPath,
-                    languageProfileId,
-                    seasons: seasons.map((s) => ({
-                      seasonNumber: s.seasonNumber,
-                      monitored: seasonMonitored[s.seasonNumber] ?? false,
-                    })),
-                  });
-                }
-              }}
-              disabled={!canSubmit}
-            >
-              {success ? (
-                <>
-                  <CheckCircle2 className="h-4 w-4 mr-1.5" />
-                  Series Added
-                </>
-              ) : addSeries.isPending ? (
-                <>
-                  <RefreshCw className="h-4 w-4 mr-1.5 animate-spin" />
-                  Adding...
-                </>
-              ) : (
-                'Request'
-              )}
-            </Button>
+            </div>
+            <div className="max-h-48 overflow-y-auto space-y-1 rounded-md border p-2">
+              {seasons.map((s) => (
+                <Label
+                  key={s.seasonNumber}
+                  className="flex items-center gap-2 text-sm cursor-pointer font-normal"
+                >
+                  <input
+                    type="checkbox"
+                    checked={seasonMonitored[s.seasonNumber] ?? false}
+                    onChange={(e) => {
+                      setSeasonMonitored((prev) => ({
+                        ...prev,
+                        [s.seasonNumber]: e.target.checked,
+                      }));
+                    }}
+                    disabled={addSeries.isPending || success}
+                  />
+                  {s.seasonNumber === 0 ? 'Specials' : `Season ${s.seasonNumber}`}
+                  {s.firstAirDate && (
+                    <span className="text-muted-foreground">— {s.firstAirDate.slice(0, 4)}</span>
+                  )}
+                </Label>
+              ))}
+            </div>
           </div>
-        </div>
-      </DialogContent>
-    </Dialog>
+        )}
+      </>
+    );
+
+  return (
+    <RequestDialog
+      open={open}
+      onClose={handleClose}
+      title="Request Series"
+      description={`${title} (${year})`}
+      isLoading={isDataLoading}
+      error={error}
+      canSubmit={canSubmit}
+      isPending={addSeries.isPending}
+      isSuccess={success}
+      successLabel="Series Added"
+      onSubmit={handleSubmit}
+    >
+      {formContent}
+    </RequestDialog>
   );
 }
