+++ packages/app-media/src/components/search/MovieSearchResult.tsx
@@ -10,6 +10,8 @@ import { highlightMatch, SearchResultItem } from '@pops/ui';
  */
 import type { ResultComponentProps } from '@pops/navigation';
 
+import { formatRuntime } from '../../lib/format';
+
 interface MovieHitData {
   title: string;
   year: string | null;
@@ -29,12 +31,9 @@ function Rating({ value }: { value: number | null }) {
   );
 }
 
-function formatRuntime(minutes: number | null): string | null {
+function formatRuntimeOrNull(minutes: number | null): string | null {
   if (minutes == null || minutes <= 0) return null;
-  const h = Math.floor(minutes / 60);
-  const m = minutes % 60;
-  if (h === 0) return `${m}m`;
-  return m > 0 ? `${h}h ${m}m` : `${h}h`;
+  return formatRuntime(minutes);
 }
 
 export function MovieSearchResult({ data }: ResultComponentProps) {
@@ -45,7 +44,7 @@ export function MovieSearchResult({ data }: ResultComponentProps) {
   const { title, year, posterUrl, voteAverage, runtime } = hit;
   const query = hit._query ?? '';
   const matchType = hit._matchType ?? 'contains';
-  const runtimeLabel = formatRuntime(runtime ?? null);
+  const runtimeLabel = formatRuntimeOrNull(runtime ?? null);
 
   return (
     <SearchResultItem
