+++ packages/app-media/src/lib/utils.ts
@@ -1,7 +1,2 @@
 export { cn } from '@pops/ui';
 
-export function formatRuntime(minutes: number): string {
-  const h = Math.floor(minutes / 60);
-  const m = minutes % 60;
-  return h > 0 ? `${h}h ${m}m` : `${m}m`;
-}
