+++ packages/navigation/package.json
@@ -14,8 +14,12 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@pops/api-client": "workspace:*",
+    "@pops/ui": "workspace:*",
+    "lucide-react": "^1.8.0",
     "react": "^19.2.5",
-    "react-router": "^7.14.1"
+    "react-router": "^7.14.1",
+    "zustand": "^5.0.5"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.9.1",
